requirements.txt
discord.py
pytubefix
ffmpeg
python-dotenv
asyncio 